requirements.txt
# Crypto Trader Pro - Requirements
# Professional cryptocurrency trading bot dependencies
# Python 3.8+ compatible

# ================================================================
# CORE DEPENDENCIES (Required)
# ================================================================

# Cryptocurrency exchange integration
ccxt>=4.2.0,<5.0.0

# Data manipulation and scheduling
pandas>=2.1.0,<3.0.0
schedule>=1.2.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# System monitoring
psutil>=5.9.0,<6.0.0

# Terminal output formatting
colorama>=0.4.6,<1.0.0

# Testing framework
pytest>=7.4.0,<8.0.0

# ================================================================
# WEB DASHBOARD DEPENDENCIES
# ================================================================

# Web dashboard framework
streamlit>=1.28.0,<2.0.0

# Interactive data visualization
plotly>=5.17.0,<6.0.0

# Streamlit-optimized statistical visualization
altair>=5.1.0,<6.0.0

# Auto-refresh functionality for real-time dashboards
streamlit-autorefresh>=1.0.1,<2.0.0

# Advanced chart components for Streamlit
streamlit-echarts>=0.4.0,<1.0.0

# ================================================================
# OPTIONAL DEPENDENCIES
# ================================================================

# Technical analysis (future use)
# ta>=0.10.2,<1.0.0

# Data visualization (future use)
# matplotlib>=3.7.0,<4.0.0

# Telegram notifications (enabled for Phase 4)
aiohttp>=3.9.0,<4.0.0

# Enhanced logging (optional)
# loguru>=0.7.0,<1.0.0

# ================================================================
# PHASE 4 DEPENDENCIES (Notifications & Backup)
# ================================================================

# Database ORM and encryption
sqlalchemy>=2.0.0,<3.0.0
bcrypt>=4.0.0,<5.0.0
pyjwt>=2.8.0,<3.0.0
cryptography>=41.0.0,<42.0.0

# Backup and compression
schedule>=1.2.0

# ================================================================
# INSTALLATION
# ================================================================
#
# Basic installation:
#   pip install -r requirements.txt
#
# Development setup:
#   pip install -r requirements.txt
#   pip install pytest black flake8
#
# ================================================================
# NOTES
# ================================================================
#
# - All core functionality uses Python standard library where possible
# - Optional dependencies can be installed as needed
# - Tested on Python 3.8, 3.9, 3.10, 3.11, 3.12
# - ccxt provides all exchange API integration
# - SQLite database requires no additional installation